pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quantum-crypto-lab"
version = "0.1.0"
description = "量子暗号プロトコル実験のためのラボ（Shorアルゴリズム・QKD等）"
requires-python = ">=3.10"
dependencies = [
  "qiskit==1.2.4",
  "matplotlib>=3.9",
  "numpy>=1.26",
  "pandas>=2.2",
  "jupyterlab>=4.2",
  "typer>=0.9.0",
]

[project.optional-dependencies]
aws = [
  "amazon-braket-sdk>=1.82",
  "qiskit-braket-provider>=0.4",
]
ibm = [
  "qiskit-ibm-runtime>=0.20",
]
hardware = [
  "amazon-braket-sdk>=1.82",
  "qiskit-braket-provider>=0.4",
  "qiskit-ibm-runtime>=0.20",
]
pqc = [
  "liboqs-python>=0.10.0",
  "cryptography>=43.0",
]
all = [
  "amazon-braket-sdk>=1.82",
  "qiskit-braket-provider>=0.4",
  "qiskit-ibm-runtime>=0.20",
  "liboqs-python>=0.10.0",
  "cryptography>=43.0",
  "qiskit-aer>=0.15.0",
]

[tool.setuptools.packages.find]
where = ["src"]
